### 2023-03-12 #5248 @editorialbot
[PRE REVIEW]: Scan Session Tool: (f)MRI scan session documentation and archiving 
**Submitting author:** @fladd (<a href="http://orcid.org/0000-0002-2754-3692">Florian Krause</a>)
**Repository:** https://github.com/fladd/ScanSessionTool
**Branch with paper.md** (empty if default branch):
**Version:** v0.9.3
