### 2023-03-29 #5248 @fladd
@cMadan Some potential reviewers I can think of:

**Stefan Heunis** (on list, 'jsheunis' on GitHub) - _experience with (real-time) (f)MRI data collection and standardization (e.g. metadata, preprocessing)_
**David Mehler** (not on list, [Google Scholar profile](https://scholar.google.com/citations?user=6nuhn44AAAAJ)) - _experience with real-time fMRI data collection and Open Science practices_
**Judith Eck** (not on list, 'ju-ec' on GitHub, [Google Scholar profile](https://scholar.google.com.sg/citations?user=mMCZmiAAAAAJ&hl=en)) - _hands-on experience with a previous version of the software to be reviewed (around 2016/2017) as a user_
