### 2023-03-31 #5248 @neuroccino
hi Chris,

thanks for reaching out, I am on holiday now and will not get to this in
the next 3-4 weeks.

David Mehler, M.D., Ph.D.
Head of *A*pplied *C*omputational *N*euroscience (*ACN*) group
Lab website: https://acn-lab.owlstown.net/
Department of Psychiatry, Psychotherapy and Psychosomatics
Uniklinik RWTH Aachen
Germany

On Fri, Mar 31, 2023 at 4:33 AM Christopher Madan ***@***.***>
wrote:

> @jsheunis <https://github.com/jsheunis> @neuroccino
> <https://github.com/neuroccino> @ju-ec <https://github.com/ju-ec>, would
> you be available and interested in reviewing this submission?
>
