### 2023-09-05 #5248 @ManavalanG
@cMadan I can review this submission if needed. I don't have experience in the field of (f)MRI, but the tool appears to be generic enough for me to review :)
