### 2023-12-06 #6126 @editorialbot
[REVIEW]: Scan Session Tool: (f)MRI scan session documentation and archiving 
**Submitting author:** @fladd (<a href="http://orcid.org/0000-0002-2754-3692">Florian Krause</a>)
**Repository:** https://github.com/fladd/ScanSessionTool
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
