### 2024-03-14 fladd/ScanSessionTool#7 @jsheunis
JOSS Paper comments
_Ping: https://github.com/openjournals/joss-reviews/issues/6126_

On the whole, the paper is clearly written and explains the purpose and capability of the software tool concisely and intuitively.

Two comments:
- there is a typo in the last paragraph, line 40:
   > preporocessing
- the paper does not cover the "state of the field". From the reviewer checklist:
   > State of the field: Do the authors describe how this software compares to other commonly-used packages?
   A couple of sentences should be added to the paper to describe existing packages and how their functionality relate to the tool being described. I know, for example, of tools like [heudiconv](https://github.com/nipy/heudiconv) and the extended set of tools in the [ReproNim](https://github.com/ReproNim) world that operate in the same space. There are likely others too.
