### 2024-03-14 fladd/ScanSessionTool#8 @jsheunis
Docs typo
_Ping: https://github.com/openjournals/joss-reviews/issues/6126_

There's a typo in the docs, under `Documentation/Usage`:

> The user interface is organized into three different areas, each hol- ding
