### 2024-03-14 fladd/ScanSessionTool#8 @jsheunis
And another one under `Documentation/Usage/The "Documents" area`

> The (f)MRI screening **from** provided by the scanning institution
