### 2024-03-14 fladd/ScanSessionTool#8 @jsheunis
And in `Documentation/Usage/The "Measurements" area`

> Click- ing on "Add Measurement" will create

The random word splitting appears elsewhere too, so I guess this text was copied from a place where the word splitting was necessaray. But here it isn't.
