### 2024-03-14 fladd/ScanSessionTool#9 @jsheunis
Statement of need in docs
Ping: https://github.com/openjournals/joss-reviews/issues/6126

The JOSS checklist asks:

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

This is not available in the documentation nor in the github repo itself.
