### 2024-03-14 fladd/ScanSessionTool#10 @jsheunis
Installation process
Ping: https://github.com/openjournals/joss-reviews/issues/6126

The JOSS checklist asks:

> Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

Installation instructions are clear, although I am unsure why `pipx` is used over `pip`, can you clarify?

Also, there are no explicitly stated dependencies. If there are none, it would would be good to state that.
