### 2024-03-14 fladd/ScanSessionTool#11 @jsheunis
Example usage
Ping: https://github.com/openjournals/joss-reviews/issues/6126

The JOSS checklist asks:

> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

I think it would be good to add an example to the docs, perhaps in the form of a walk-through tutorial. This would make adoption for users much easier.
