### 2024-03-14 fladd/ScanSessionTool#7 @fladd
Thanks!

The "state of the field" requirement is an interesting one, as this is not part of the author checklist :-) I think this is also not covered in their example paper in the documentation: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography.

But I am of course happy to include a comparison to the two packages you mentioned. Finding related work was quite difficult (we tried before) actually, since we seem to use a different terminology (i.e. searching for "MRI scan session documentation" didn't really result in anything useful).
