### 2024-03-14 fladd/ScanSessionTool#9 @fladd
Yes, indeed. This was not a requirement in the author documentation. I will add this.
