### 2024-03-14 fladd/ScanSessionTool#10 @fladd
Pipx mainly because it is an application only and not also a library, and it makes thing much easier for end users (no venv creation needed, readily available system command to run it, etc.). But using pip will work equally of course for those who prefer that.
Would you suggest to mention this explicitly?

Dependencies are pyyaml and pydicom, and that is indeed not stated anywhere (mainly, because the end user doesn't care, since pipx will take care of those automatically during installation). So from an end-user perspective, there are no dependencies that need fulfilling (on their side). I can be more explicit about this though and state that there are technical dependencies, but that the user does not need to take care of them themselves?
