### 2024-03-14 fladd/ScanSessionTool#11 @fladd
Thanks for the comment. I will think about how to best implement this.
