### 2024-03-14 fladd/ScanSessionTool#7 @jsheunis
Interesting. Good feedback for JOSS then ;) However, I think it's definitely a good idea to include something like "state of the field" in such a paper, as it provides useful context for the reader or potential user.

Re the specific tools I shared, I only know of them through my own involvement in related projects, and their popularity in the open neuroimaging community, but I'm sure others exist. Just as I'm sure it will be difficult to find them :)
