### 2024-03-14 fladd/ScanSessionTool#10 @jsheunis
Thanks for the explanation. This makes sense.

I'm always a fan of making things explicit.

What i think could be a good idea is to create some separation between user docs and developer/contributor docs. I.e. for users, a more minimal set of instructions and dependencies can be provided, with the focus being getting the tool installed on their system and using it is as easily as possible.

For developers I think stating requirements explicitly (e.g. in `requirements.txt`), being able to install the package in a virtual environment, and making the package available via pip would be very beneficial. This relates to another issue: https://github.com/fladd/ScanSessionTool/issues/12
