### 2024-03-14 fladd/ScanSessionTool#12 @jsheunis
Contributing guidelines
Ping: https://github.com/openjournals/joss-reviews/issues/6126

The checklist asks:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

This is mostly absent from all parts of the project, and I think this should definitely be addressed in order to make the project more contributor-friendly. A common approach is to have a `CONTRIBUTING` file in the repo's root directory that describes everything anyone would need to know in order to contribute, from creating issues to contributing code (and how to test their code contributions). The same content, or a slight variation of it, could be added to the user documentation, e.g. under a "For developers" section.
