### 2024-03-14 fladd/ScanSessionTool#10 @fladd
Yes, I can add the requirements.txt for developers to the repo.

Concerning installation via pip: That should be entirely possible. The package is on PyPi (that is also where pipx gets it from). Did you run into any issues installing with pip?
