### 2024-03-14 fladd/ScanSessionTool#12 @fladd
Thanks for the suggestion. Can you maybe point me to an example for this file? We are not doing anything special with respect to contributions (i.e. we do accept pull requests and issues on GitHub), so I am not entirely sure what should go in there exactly. It seems that in our case, such a file could just point to the GitHub documentation on [creating issues](https://docs.github.com/en/issues/tracking-your-work-with-issues/creating-an-issue) and [creating pull requests](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/creating-a-pull-request) (they explain this very well I think, and I am not sure what I could do better).
