### 2024-03-14 fladd/ScanSessionTool#13 @jsheunis
Testing
_Ping: https://github.com/openjournals/joss-reviews/issues/6126_

Some notes about testing:

1. I wanted to run/conduct any tests that the package includes, so I went to the location that seemed most intuitive for me ==> `./tests/`. There I found a readme that described the test setup. But this is not a foolproof way of telling people about testing. It could be made easier for developers/users to know what to do if they want to run tests. Especially developers need to know how to run tests if they're contributing code, so you can e.g. include a few notes on testing in the dev docs mentioned in https://github.com/fladd/ScanSessionTool/issues/12.
2. When following the testing steps, it's not obvious where exactly the protocol file (`ScanProtocol_TestData_sub-001_ses-007-Transfer_20211203.txt`) should be found. E.g. when I clicked "Open", the file browser on my system opened in a random location. I wasn't sure if the tool should automatically open the file browser in the package installation location, or whether the testing data was actually even installed together with the package. So I just cloned the git repo and navigated to the test protocol file in the local clone. But this wouldn't be obvious to someone who is unfamiliar with these systems. So some instructions would be good to improve user experience here.
3. The testing pipeline described in the mentioned readme feels almost more like a tutorial to me. I think this, or an adaptation of it, could actually be appropriate in the context of the walkthrough tutorial mentioned in https://github.com/fladd/ScanSessionTool/issues/11
4. I read the part about "Since Scan Session Tool is a graphical application, these tests have to be performed manually.", but I don't think the fact that it's a graphical application prevents the use of automated tests. I actually think the same tests can be automated, just without the clicking and visual verifying bits although these can be replaced with mock functions. And having automated tests (run via e.g. `pytest`) will allow: a) developers to run local tests to test their code contributions against, and b) you to set up continuous integration (via e.g. appveyor and/or github actions) so that tests can be run against new PRs. Both of these additions, automated testing and continuous integration, would improve the testing setup and the package itself substantially.
