### 2024-03-14 fladd/ScanSessionTool#12 @jsheunis
Sure, here are a few examples, with varying detail/complexity/approaches:

- Numpy: https://github.com/numpy/numpy?tab=readme-ov-file#call-for-contributions
- datalad: https://github.com/datalad/datalad/blob/maint/CONTRIBUTING.md
- nilearn: https://github.com/nilearn/nilearn/blob/main/CONTRIBUTING.rst and https://nilearn.github.io/stable/development.html#contributing
- datalad-catalog: https://github.com/datalad/datalad-catalog#7-contributing
- scikit-image: https://scikit-image.org/docs/stable/development/contribute.html

The github resources are definitely useful, but they don't convey the complete picture of what contributing could mean, and what form it could take.
