### 2024-03-14 fladd/ScanSessionTool#10 @jsheunis
I didn't try installing via pip before making that comment. I did now, and all works well 👍
