### 2024-03-14 fladd/ScanSessionTool#14 @jsheunis
Code organization
_Ping: https://github.com/openjournals/joss-reviews/issues/6126_

To improve code organization and readability, I think it could help to split the current `scansessiontool.py` module into mutliple modules. At the moment it is about 2.4k lines in a single file. I scanned through the code and it seems there are already a few natural candidates for separate modules, e.g. some utility functions, the dialogue classes, the app class, etc.
