### 2024-03-14 #6126 @jsheunis
I went through all checklist items and created relevant issues. I can check off the remaining open boxes once those issues are addressed. There's one remaining point that I would like to comment on, and that is the question of "Substantial scholarly effort: Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)"

I find it difficult to pass clear judgement on this item, i.e. to give a YES/NO answer. Looking at the specific items mentioned in the JOSS guidelines:

> Age of software (is this a well-established software project) / length of commit history

This is fine, the tool is several years in development

> Number of commits.

No problem here.

> Number of authors.

Might be slightly concerning because it's two authors, but actually is not a problem in and of itself.

> Total lines of code (LOC). Submissions under 1000 LOC will usually be flagged, those under 300 LOC will be desk rejected.

No specific reason for concern

> Whether the software has already been cited in academic papers.

It has, but by the authors themselves in their other work.

> Whether the software is sufficiently useful that it is likely to be cited by your peer group.

I am uncertain about this. The main reason is that ScanSessionTool seems like a useful package (no doubt) developed by _and for_ the authors and their close collaborators. There isn't clear evidence to me that the package has taken hold in the wider open source community or that it will necessarily expand much beyond the existing user group. In my experience, each MRI data collection site or group has their own approach to organising DICOM data (which might be problematic or extremely streamlined/intuitive, or somewhere in between), and there is no standard way of doing this. ScanSessionTool proposes a standard hierarchy (if I understood the documentation correctly?), in essence implying the adoption of a data organization standard and not merely the use of a tool. The adoption of a standard is a much greater challenge than using a tool built around a specific standard that has already been adopted. My suspicion is that groups might choose to not use the tool because they do not want to change their standard approach (because they have status quo bias or changing will affect to many downstream processes or whatever reason).

My viewpoint is also somewhat related to the JOSS point that
> software should be feature-complete (i.e., no half-baked solutions), packaged appropriately according to common community standards [...], and designed for maintainable extension. “Minor utility” packages, including “thin” API clients, and single-function packages are not acceptable.

It would not be fair to call ScanSessionTool a minor utility package (and I'm not doing that) but some aspects of its current implementation do lean towards that direction. I'm thinking specifically about the 2.4k-line single module, the absence of unit-tests and continuous integration, and the package's (currently) limited contributor community. However, these can all be addressed.

I'm happy to receive arguments/evidence against my point of view. In any event, I don't think my being on the fence on this particular point should constitute any form of rejection (in the traditional journal-centric sense of the word). IMO this is a useful package that people should know about and can benefit from. I would appreciate thoughts on this from other the reviewer(s), the editor, and authors.
