### 2024-03-19 fladd/ScanSessionTool#12 @fladd
Thanks, we will implement this.
