### 2024-03-19 fladd/ScanSessionTool#13 @fladd
Thanks for the suggestions.

1. We will include notes on testing in the dev docs.
2. We will be explicit that one has to navigate to the protocol file.
3. We did think about unit test intensively in the past and decided back then to not do it, due to all the difficulties involved in testing GUI functionality automatically. However, given your suggestion, we reviewed the possibilities again, and we will automatize part of our testing, where splitting logic from GUI is appropriate. In particular, we will automatize the archiving procedure (that is, the second part of the current testing procedure: loading of a protocol file and archiving based on the information in that). We decided, however, not to test the interaction of each element in the GUI (that is, the first part of the current testing procedure: loading/saving protocol files to/from information stored in GUI elements), as this functionality is inherently bound to user interaction.
