### 2024-03-19 fladd/ScanSessionTool#14 @fladd
Thanks for the suggestion. We will modularize the code as suggested.
