### 2024-03-19 #6126 @fladd
Thanks for reviewing and for your suggestions and opened issues!

Concerning your concern on substantial scholarly effort:
We of course went through the checklist in detail before making the decision to submit to JOSS, and we were confident that we fulfill all of the requirements. You seem to also agree with this in general, but have a slight concern about one point specifically ("Whether the software is sufficiently useful that it is likely to be cited by your peer group.").

>I am uncertain about this. The main reason is that ScanSessionTool seems like a useful package (no doubt) developed by and for the authors and their close collaborators. There isn't clear evidence to me that the package has taken hold in the wider open source community or that it will necessarily expand much beyond the existing user group. In my experience, each MRI data collection site or group has their own approach to organising DICOM data (which might be problematic or extremely streamlined/intuitive, or somewhere in between), and there is no standard way of doing this. ScanSessionTool proposes a standard hierarchy (if I understood the documentation correctly?), in essence implying the adoption of a data organization standard and not merely the use of a tool. The adoption of a standard is a much greater challenge than using a tool built around a specific standard that has already been adopted. My suspicion is that groups might choose to not use the tool because they do not want to change their standard approach (because they have status quo bias or changing will affect to many downstream processes or whatever reason).

While it is true that we initially created this software for our own usage (which is probably the case for most research software), we don't see any reason why it shouldn't be useful to others as well. Not only is the scan session documentation and protocol logging part useful in itself (i.e. without the archiving), but I also don't think that the archiving hierarchy we use here has much impact on downstream pipelines, given that most of these pipelines nowadays resolve around the BIDS format, to which our structure can be converted to easily (e.g. using BIDS Coiner, as mentioned in the paper).

When thinking about the perceived lack of current adoption of the tool, there are multiple things to consider:
1. Scholars tend to cite software only when they can refer to a journal article (hence this submission).
2. Accurately tracking usage in ways other than citations is rather difficult. We don't do any web traffic analytics ourselves, and the potentially best indicator might be the number of downloads from PiPy, which show >20 downloads per month.
3. While the package is publicly available on GithHub, we did not put any effort into actively advertising it. A publication will also help with visibility.

> It would not be fair to call ScanSessionTool a minor utility package (and I'm not doing that) but some aspects of its current implementation do lean towards that direction. I'm thinking specifically about the 2.4k-line single module, the absence of unit-tests and continuous integration, and the package's (currently) limited contributor community. However, these can all be addressed.

We are happy to address all of these points in due time.
