### 2024-03-19 #6126 @fladd
Thanks for the reply @neuroccino!

1. Good point, we will add a sentence or two to the manuscript to clarify what this special support is about. (It is already explained in detail in the documentation).
2. There is a link to "Repository" in the left column in the paper (autogenerated). The content of the "Name" dropdown menu is fully modifiable on a per-project basis, with different sets of names possible for each of the three available "Types" (anat, func, misc). All of this can be customized as part of study-specific templates in a YAML configuration file (see [here](https://github.com/fladd/ScanSessionTool/blob/master/sst.yaml) for an example). (Of course, it is also possible to just type in a name that is not one of the predefined options in the dropdown menu.)
