### 2024-04-10 ff4b444
Code reorganization
Implements #14
