+++ scansessiontool/scansessiontool.py
@@ -14,7 +14,6 @@ import json
 import shutil
 import threading
 import multiprocessing
-from tempfile import mkstemp
 
 if sys.version[0] == '3':
     from tkinter import *
@@ -32,569 +31,24 @@ else:
     from ScrolledText import ScrolledText
 
 import yaml
-import pydicom
 
 from .__meta__ import __version__, __date__
+from .widgets import (FixedSizeFrame,
+                      AutoScrollbarText,
+                      VerticalScrolledFrame,
+                      AutocompleteCombobox,
+                      Spinbox)
+from .dialogues import (ArchiveDialogue,
+                        BusyDialogue,
+                        MessageDialogue,
+                        HelpDialogue)
+from .utilities import (replace,
+                        readdicom)
+
+
+class ScanSessionTool(Frame):
+    """The main Scan Session Tool Tkinter application."""
 
-
-docs = """
-+ - - - - - - - - - - - - - - Scan Session Tool - - - - - - - - - - - - - - +
-|                                                                           |
-|        A tool for MR scan session documentation and data archiving        |
-|                                                                           |
-|             Authors: Florian Krause <f.krause@donders.ru.nl>              |
-|                      Nikos Kogias <n.kogias@student.ru.nl>                |
-|                                                                           |
-+ - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - +
-
-
-
-=================================== About ===================================
-
-The Scan Session Tool is a graphical application for documenting (f)MRI scan
-sessions and automatized data archiving. Information about the scan session
-itself, used forms and documents, as well as the single measurements can be
-entered and saved into a protocol file. This information can furthermore be
-used to copy acquired data (DICOM images as well as optional stimulation
-protocols and logfiles into a specific hierarchical folder structure for
-unified archiving purposes, with optional sepcial support for
-(Turbo-)BrainVoyager (https://brainvoyager.com).
-
-
-
-=================================== Usage ===================================
-
-The user interface is organized into three different content areas, each hol-
-ding different information about the scan session, as well as an additional
-control area for opening and saving session information and for automatically
-archiving acquired data, based on the session information.
-
-
------------------------ The "General Information" area ----------------------
-
-This area provides input fields for basic information about the scan session.
-Some of the fields allow for a selection of pre-specified values taken from
-a config file (see below), while others take freely typed characters. Fields
-that are marked with a red background, are mandatory and need to be filled
-in. Fields that are marked with an orange background are automatically filled
-in, but need to be checked.
-The following fields are available:
-    "Project"        - The project identifier
-                       (free-type and selection)
-    "Subject"        - The subject number
-                       (001-999)
-                     - The subject type
-                       (free-type and selection)
-    "Session"        - The session number
-                       (001-999)
-                     - The session type
-                       (free-type and selection)
-    "Date"           - The date of the scan session
-                       (free-type, auto-filled)
-    "Time A"         - The main time period (e.g. official scanner booking)
-                       (free-type)
-    "Time B"         - An additional time period (e.g. actual scanner usage)
-                       (free-type)
-    "User 1"         - The main user (e.g. responsible MR operator)
-                       (free-type and selection)
-    "User 2"         - An additional user (e.g. back-up/buddy)
-                       (free-type and selection)
-    "Notes"          - Any additional notes about the session
-                       (free-type)
-
-
[…]
