+++ scansessiontool/widgets.py
@@ -1,0 +1,289 @@
+"""Widgets.
+
+A set of custom Tkinter widgets used in Scan Session Tool.
+
+"""
+
+
+import os
+import sys
+import glob
+import shutil
+import platform
+
+if sys.version[0] == '3':
+    from tkinter import *
+    from tkinter.ttk import *
+else:
+    from Tkinter import *
+    from ttk import *
+
+class FixedSizeFrame(Frame):
+    """A Tkinter frame with a fixed size."""
+
+    def __init__(self, master, width, height, **kw):
+        Frame.__init__(self, master, **kw)
+        self["width"] = width
+        self["height"] = height
+        self.grid_propagate(False)
+        self.grid_rowconfigure(0, weight=1)
+        self.grid_columnconfigure(0, weight=1)
+
+
+class AutoScrollbarText(Text):
+    """A Tkinter text widget with an automatically showing/hiding scrollbar."""
+
+    def __init__(self, master=None, **kw):
+        self.frame = Frame(master)
+        self.yvbar = Scrollbar(self.frame)
+        self.yvbar.grid(row=0, column=1, sticky="NS")
+
+        kw.update({'yscrollcommand': self.set_yvbar})
+        Text.__init__(self, self.frame, undo=True, **kw)
+        self.grid(row=0, column=0, sticky="WENS")
+        self.frame.grid_rowconfigure(0, weight=1)
+        self.frame.grid_columnconfigure(0, weight=1)
+        self.yvbar['command'] = self.yview
+
+        self.undo_history = 0
+        if platform.system() == "Darwin":
+            self.bind("<Command-z>", self.undo_dummy)
+            self.bind("<Command-Z>", self.redo)
+        else:
+            self.bind("<Control-z>", self.undo_dummy)
+            self.bind("<Control-Z>", self.redo)
+
+        # Copy geometry methods of self.frame without overriding Text
+        # methods -- hack!
+        text_meths = list(vars(Text).keys())
+        methods = list(vars(Pack).keys()) + list(vars(Grid).keys()) + \
+                list(vars(Place).keys())
+        methods = set(methods).difference(text_meths)
+
+        for m in methods:
+            if m[0] != '_' and m != 'config' and m != 'configure':
+                setattr(self, m, getattr(self.frame, m))
+
+    def undo_dummy(self, *args):
+        if self.edit_modified():
+            self.undo_history += 1
+
+    def redo(self, *args):
+        if self.undo_history > 0:
+            self.edit_redo()
+            self.undo_history -= 1
+
+    def set_yvbar(self, lo, hi):
+        if float(lo) <= 0.0 and float(hi) >= 1.0:
+            # grid_remove is currently missing from Tkinter!
+            self.yvbar.tk.call("grid", "remove", self.yvbar)
+        else:
+            self.yvbar.grid()
+        self.yvbar.set(lo, hi)
+
+    def __str__(self):
+        return str(self.frame)
+
+    def copy_logfiles(self, source, destination):
+        original = self.get(1.0, END)
+        logfiles = original.split("\n")
+        logfiles = [x.strip() for x in logfiles if x != ""]
+        warning = ""
+        new = original
+        for logfile in logfiles:
+            if "*" in logfile:
+                replaced = []
+            try:
+                if logfile != "" and not os.path.isdir(logfile):
+                    files = glob.glob(os.path.join(source, logfile))
+                    if files == []:
+                        raise Exception
+                    for file_ in files:
+                        if not os.path.isdir(file_):
+                            if "*" in logfile:
+                                replaced.append(
+                                    os.path.split(file_)[-1])
+                            shutil.copyfile(
+                                file_,
+                                os.path.join(destination,
+                                             os.path.split(file_)[-1]))
[…]

[…]
