+++ README.md
@@ -41,3 +41,28 @@ Documentation
 The full documentation can be found from within the programme, by clicking on
 the "?" button, or by selecting "Scan Session Tool Help" from the Help menu,
 as well as at https://fladd.github.io/ScanSessionTool/.
+
+Contributing
+------------
+We very much welcome contributions to Scan Session Tool, and there are multiple ways to contribute. Contributions are always handled transparently through the public GitHub interface.
+
+### Bug reports
+If you encounter bugs or incorrect/unexpected behaviour of Scan Session Tool, please consider creating an [issue](https://github.com/fladd/ScanSessionTool/issues) here on GitHub for it. You can read more about how to do this [here](https://docs.github.com/en/issues/tracking-your-work-with-issues/creating-an-issue). When reporting issues, make sure to report (1) what you were trying to do, (2) what the expected outcome should have been, (3) what the observed outcome is, and (4) how to replicate this behaviour.
+
+### Code contributions
+We work with the standard [GitHub forking workflow](https://docs.github.com/en/get-started/exploring-projects-on-github/contributing-to-a-project) and appreciate code contributions in that way. Individuals that have contributed code will be named with their contributions on this page.
+
+#### Bug fixes
+To implement bug fixes, please
+1. [fork](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/working-with-forks/fork-a-repo) the [main branch](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/about-branches) of this repository
+2. [commit](https://docs.github.com/en/pull-requests/committing-changes-to-your-project) your changes (potentially in a [new branch](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/about-branches))
+3. [test](https://github.com/fladd/ScanSessionTool/tree/master/tests) you changes
+4. [create a pull request](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/creating-a-pull-request) and reference the corresponding GitHub [issue](https://github.com/fladd/ScanSessionTool/issues), if applicable
+   
+#### New features
+Before implementing a new feature, we would encourage you to first discuss the planned feature with us, either in the [issue tracker](https://github.com/fladd/ScanSessionTool/issues), or [GitHub discussions](https://github.com/fladd/ScanSessionTool/discussions). Once discussed, new features can then be implemented the same way as bug fixes.
+
+Contributors
+------------
+- Florian Krause (main developer)
+- Nikos Kogias (co-developer)
