### 2024-04-10 e9965fa
Added statement of need in docs
Implementing #9
