+++ scansessiontool/help.txt
@@ -11,6 +11,19 @@
 
 =================================== About ===================================
 
+There is an urgent need to improve the reproducibility of (functional) MRI
+research through transparent reporting, but standardization in this domain is
+currently lacking. Shared MR images are often only made available after
+transformation into a derivative data format (e.g. BIDS;
+https://bids.neuroimaging.io/), and scan session documentation is commonly
+either manually implemented (e.g. with hand written notes) or neglected
+entirely.
+
+Scan Session Tool was written to fill this gap, and to be used by
+neuroscientists, to help them increase transparency and reproducibilty of
+their MRI research by standardizing scan session documentation and raw data
+archiving.
+
 The Scan Session Tool is a graphical application for documenting (f)MRI scan
 sessions and automatized data archiving. Information about the scan session
 itself, used forms and documents, as well as the single measurements can be
