### 2024-04-10 fladd/ScanSessionTool#8 @fladd
Fixed in https://github.com/fladd/ScanSessionTool/commit/4c07d8cc9f8e7a358db98ff525f493f120549e92 and https://github.com/fladd/ScanSessionTool/commit/eb270ee63e5a612ac464a3bb2c14a939219b1279
