### 2024-04-10 d78c461
Updated installation instructions
Implementing #10
