### 2024-04-14 4185789
Added automated tests
Implementing #13
