### 2024-04-14 fladd/ScanSessionTool#13 @fladd
After having spent (way too) many hours on this now, I eventually did add tests for both scan session protocol handling and data archiving! See main README.md for simple instructions on how to run.

**Note:**
Due to the way Scan Session Tool works and is implemented, I decided against separating the application logic from the GUI and testing it in isolation (because part of the core functionality of the proramme is closely tied to what can be seen on the screen; i.e. the scan session documentation). Running unit tests on GUI applications is not trivial and requires hacks to somehow intercept events into the main event loop from the outside. Common solutions for this (such as multithreading and asynchronious code (e.g. https://stackoverflow.com/questions/42295186/unittesting-a-tkinter-application, https://stackoverflow.com/questions/4083796/how-do-i-run-unittest-on-a-tkinter-app) did not work in this case. In the end, I had to implement part of the testing logic in the application itself, such that it can be triggered internally, based on how the application is started. This seems to work rather well I think!

I might, in the future, implement CI for this as well (have to read more into dummy displays for headless testing (e.g. https://arbitrary-but-fixed.net/2022/01/21/headless-gui-github-actions.html).
