### 2024-04-17 b63c49a
test fixes for Windows
