+++ paper/paper.bib
@@ -214,3 +214,41 @@ Publisher: Nature Publishing Group},
 	pmcid = {PMC4978148},
 	pages = {160044},
 }
+
+@software{halchenko_2024_11201247,
+  author       = {Halchenko, Yaroslav O. and
+                  Goncalves, Mathias and
+                  Ghosh, Satrajit and
+                  Velasco, Pablo and
+                  Visconti di Oleggio Castello, Matteo and
+                  Salo, Taylor and
+                  Wodder II, John T. and
+                  Hanke, Michael and
+                  Sadil, Patrick and
+                  Gorgolewski, Krzysztof Jacek and
+                  Ioanas, Horea-Ioan and
+                  Rorden, Chris and
+                  Hendrickson, Timothy J. and
+                  Dayan, Michael and
+                  Houlihan, Sean Dae and
+                  Kent, James and
+                  Strauss, Ted and
+                  Lee, John and
+                  To, Isaac and
+                  Markiewicz, Christopher J. and
+                  Lukas, Darren and
+                  Butler, Ellyn R. and
+                  Thompson, Todd and
+                  Termenon, Maite and
+                  Smith, David V. and
+                  Macdonald, Austin and
+                  Kennedy, David N.},
+  title        = {{HeuDiConv: flexible DICOM conversion into 
+                   structured directory layouts}},
+  month        = may,
+  year         = 2024,
+  publisher    = {Zenodo},
+  version      = {v1.1.3},
+  doi          = {10.5281/zenodo.11201247},
+  url          = {https://doi.org/10.5281/zenodo.11201247}
+}
