### 2024-05-29 fladd/ScanSessionTool#11 @fladd
We have implemented this now. Thanks again for the suggestion.
