### 2024-05-29 fladd/ScanSessionTool#7 @fladd
Thanks again, we have implemented these suggestions now.
