### 2024-05-29 #6126 @fladd
Thanks for picking this up again @Kevin-Mattheus-Moerman.

I am also happy to report that we, in the meantime, have implemented all changes and suggestions raised by the reviewers. For a full overview of all changes, please see https://github.com/fladd/ScanSessionTool/pull/16 and the corresponding (linked) issues.

Many thanks to the reviewers again for their valuable input!
