### 2024-06-07 fladd/ScanSessionTool#18 @jsheunis
Tests fail locally
I have a mac and have the latest state of the `master` branch installed in a virtual environment. The tests fail when I run them:

```
> python -m unittest -v

setUpModule (tests.tests) ... ERROR
setUpModule (tests.tests) ... ERROR

ERROR: setUpModule (tests.tests)
Traceback (most recent call last):
  File "/Users/jsheunis/Documents/ScanSessionTool/tests/tests.py", line 24, in setUpModule
    DATA_DIR = tempfile.TemporaryDirectory(delete=False)
               ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
TypeError: TemporaryDirectory.__init__() got an unexpected keyword argument 'delete'

ERROR: setUpModule (tests.tests)
Traceback (most recent call last):
  File "/Users/jsheunis/Documents/ScanSessionTool/tests/tests.py", line 30, in cleanup
    DATA_DIR.cleanup()
    ^^^^^^^^^^^^^^^^
AttributeError: 'NoneType' object has no attribute 'cleanup'

Ran 0 tests in 0.022s

FAILED (errors=2)
```
