### 2024-06-07 #6126 @jsheunis
Thanks for the ping. I've gone through the list again and ticked off the parts that I could confirm have been addressed. Nice work @fladd.

The only remaining aspect is the testing. I have 2 points:
- I tried to run the tests locally and they failed, see the relevant issue: https://github.com/fladd/ScanSessionTool/issues/18
- I think it makes sense to add a simple github action to run the tests on PRs and push/merge to master.
