### 2024-06-07 fladd/ScanSessionTool#18 @fladd
Ah, sorry, the "delete" argument only works in Python 3.12, but not in previous versions. I rewrote that part now, such that it does not rely on this argument anymore. Should run now for you as well.
