### 2024-06-07 #6126 @fladd
@jsheunis It seems that the test required Python version 3.12. I rewrote it now, such that it should work with previous versions as well. I just confirmed that it now runs on MacOS with Python 3.10, so please try again. Sorry for the inconvenience.

Regarding a GitHub Action for the tests: I thought about this, too. It could be nice, but the tests require the GUI to be shown while GitHub's CI is headless of course...so this is not straight forward, if possible at all. Given that (1) the new automated tests I implemented now (which are arguably much nicer) can easily be run locally with a single command and (2) a GitHub Action is not required by JOSS (technically, even the manual testing procedure we had before would have fulfilled the requirements), I feel that the work we would need to put into figuring this out would at this point be disproportional to the rather small improvement a GitHub Action would bring over the current situation.
