### 2024-06-10 fladd/ScanSessionTool#18 @jsheunis
Can confirm that it works locally now for Python 3.11.8.
