### 2024-06-10 #6126 @jsheunis
I didn't consider the headless requirement, your argument makes sense. I would add that whether or not an automated CI setup is a requirement from JOSS or not, it's still a good idea in general. So worth keeping in mind for future, but IMO not a dealbreaker for the article itself.
