+++ scansessiontool/__meta__.py
@@ -1,2 +1,2 @@
-__version__ = '1.0.1.dev0'
+__version__ = '1.0.0.dev0'
 __date__ = '10 Jun 2024'
