### 2024-06-10 60ed459
added help.txt to package data
