### 2024-06-10 8c194a4
Fix for loading default config file
