### 2024-06-10 1a800bc
moved default config into package
