+++ setup.py
@@ -35,7 +35,7 @@ setup(
         "Operating System :: OS Independent",
     ],
     packages = ['scansessiontool'],
-    package_data = {'scansessiontool': ['help.txt']},
+    package_data = {'scansessiontool': ['help.txt', 'sst.yaml']},
     python_requires=">=3.6",
     setup_requires = ['wheel'],
     install_requires = ['pyYAML',
