### 2024-06-11 #6126 @fladd
If there are no further comments, then I would release a 1.0.0 version that incorporates all the latest changes. (A pre-release of this is already available here: https://github.com/fladd/ScanSessionTool/releases/tag/v1.0.0.dev3).
