### 2024-06-11 2000efb
added SST icon
