+++ scansessiontool/__meta__.py
@@ -1,2 +1,2 @@
-__version__ = '1.0.0.dev4'
+__version__ = '1.0.0.dev5'
 __date__ = '101 Jun 2024'
