### 2024-06-11 b6ed650
Bumped version to 1.0.0
