### 2024-06-11 d52e222
added sst_icon.png
