+++ setup.py
@@ -35,7 +35,8 @@ setup(
         "Operating System :: OS Independent",
     ],
     packages = ['scansessiontool'],
-    package_data = {'scansessiontool': ['help.txt', 'sst.yaml', 'sst_icon.ico']},
+    package_data = {'scansessiontool': ['help.txt', 'sst.yaml', 'sst_icon.ico',
+                                        'sst_icon.png']},
     python_requires=">=3.6",
     setup_requires = ['wheel'],
     install_requires = ['pyYAML',
