+++ scansessiontool/__main__.py
@@ -1,4 +1,3 @@
-import os
 import platform
 from tkinter import *
 from tkinter.ttk import *
@@ -8,13 +7,6 @@ from .scansessiontool import ScanSessionTool
 
 def run():
     root = Tk()
-    if platform.system() == "Windows":
-        root.iconbitmap(os.path.abspath(os.path.join(
-            os.path.split(__file__)[0], "sst_icon.ico")))
-    else:
-        root.tk.call('wm', 'iconphoto', root._w,
-                     PhotoImage(file=os.path.abspath(os.path.join(
-                os.path.split(__file__)[0], "sst_icon.png"))))
     app = ScanSessionTool(root)
     app.mainloop()
 
