### 2024-06-11 b602415
Moved setting app icon into scansessiontool.py
