### 2024-06-11 0d16a16
improve initial window drawing
