### 2024-06-11 23b5dcd
Version bumped to 1.0.0 (again)
