+++ scansessiontool/__meta__.py
@@ -1,2 +1,2 @@
-__version__ = '1.0.0.dev6'
+__version__ = '1.0.0'
 __date__ = '11 Jun 2024'
