### 2024-06-11 #6126 @fladd
I went ahead and released the new final version with the latest changes (plus a new application logo). I also updated Figure 1 in the paper with a screenshot showing the actual version.
