### 2024-07-18 #6126 @Kevin-Mattheus-Moerman
@fladd apologies for the delays encountered here so far. It should be smooth sailing from here on.
Please have a look at the above checklist. Can you address these points please. I've just set the version tag to be `v1.0.0`, so if that release is the one we'll go with please use that for the Zenodo archive too (be sure to include the `v` in the version tag as it should fully match the GitHub release tag).

On the paper:
- [ ] Please check typo: `reproducibilty` -> `reproducibility`
- [ ] Please check typo: `succesfully` -> `successfully`
- [ ] The paper mainly features European English, please edit the paper and fix use of American English e.g.:
* `standardized`
* `automatized` -> `automatised` but also consider `automated` instead
* `organize` -> `organise`
* `recognized` -> `recognised`
* `standardization` -> `standardisation`
