### 2024-07-19 86a1438
Update paper.md
Fixes typos and British English as suggested here: https://github.com/openjournals/joss-reviews/issues/6126#issuecomment-2236165232
