### 2024-07-19 #6126 @fladd
@Kevin-Mattheus-Moerman I archived version 1.0.0 on Zenodo. The DOI is [10.5281/zenodo.12784180](https://zenodo.org/doi/10.5281/zenodo.12784180).

I also fixed the typos and British English spelling in the paper (see https://github.com/fladd/ScanSessionTool/commit/86a143839a3f55186a6ac532db0c438664864114)
