### 2024-07-22 #6126 @Kevin-Mattheus-Moerman
@fladd the version tag listed here should match a tagged release on your repository and should match the version tag listed over at ZENODO. I saw you used the tag `Scan Session Tool 1.0.0` in your repository. If this is the preferred tag then please edit the Zenodo listed version tag to match it. Or otherwise please advise on the version tag to use (e.g. you could instead create a tagged release with `v1.0.0` as the tag and we can set that here to match).
