### 2024-07-22 #6126 @fladd
@Kevin-Mattheus-Moerman

I am a bit confused. "Scan Session Tool 1.0.0" is just the descriptive name of the GitHub release. Scan Session Tool uses the [semantic versioning scheme](https://semver.org/). The current version is "1.0.0", and the corresponding GitHub tag for the release of this version is "v1.0.0" (as is [common practice](https://stackoverflow.com/questions/37781073/why-github-suggest-prefix-your-version-names-with-the-letter-v) for GitHub tags):

![image](https://github.com/user-attachments/assets/7e9654ab-9038-4028-b035-827247507662)

On Zenodo, I matched this tag (as you instructed me to do [earlier](https://github.com/openjournals/joss-reviews/issues/6126#issuecomment-2236165232)):

![image](https://github.com/user-attachments/assets/a7f27d5c-7742-4d16-a991-b80abe60a5e0)

I'm afraid that it is not entirely clear to me what you are asking me to do now. I am happy to adapt the Zenodo page to whatever you think is appropriate. Do you want me to change the version on Zenodo from "v1.0.0" to "1.0.0"?).

Could you please clarify?
