### 2024-08-06 #6126 @Kevin-Mattheus-Moerman
@fladd I think all looks good to proceed now. I am just checking about one last thing. Currently the paper does not have an "Acknowledgements" section. This is not a required section, but if you need to acknowledge funding I suggest you add the section before we process this for final acceptance. Let me know what you decide.
