### 2024-08-06 0129bcf
Updated paper.md (added acknowledgements)
