### 2024-08-07 #6126 @Kevin-Mattheus-Moerman
@fladd congratulations on this JOSS publication! We do apologize for the much longer than usual review process.

I would like to send a special thank you to @jsheunis and @neuroccino for reviewing this work, and for @cMadan for assistance in editing as well.
